refactor: clean up compiler project split
