+++ packages/compiler-wasm/crate/src/project/css.rs
@@ -1,23 +1,29 @@
 use super::WasmCompiler;
 
+use pandacss_encoder::Atom as CoreAtom;
 use pandacss_fs::{FileSystem, PathSystem};
 use serde::Serialize as _;
+use std::collections::BTreeSet;
 use std::path::Path;
+use std::sync::Arc;
 use wasm_bindgen::prelude::*;
 
-use super::serde_types::{
-    CodegenArtifactSerde, CssOutputOptionsSerde, GenerateArtifactOptionsSerde,
-    LayerCssOptionsSerde, WriteCssResultSerde, WriteFilesResultSerde,
+use super::interop::{
+    compile_options_from_js, css_output_options_from_js, parse_required_options,
+    write_css_options_from_js, write_split_css_options_from_js,
 };
-use super::support::{
-    build_compile_output, build_layer_compile_output, build_split_css, compile_options_from_js,
-    css_output_options_from_js, generate_options, parse_required_options, to_codegen_artifact,
-    write_artifacts_options_from_js, write_css_options_from_js, write_split_css_options_from_js,
+use super::serde_types::{
+    CompileFileManifestSerde, CompileLayerRangeSerde, CompileLayerRangesSerde,
+    CompileManifestSerde, CompileOutputSerde, CssOutputOptionsSerde, LayerCssOptionsSerde,
+    SplitCssFileSerde, WriteCssResultSerde, WriteFilesResultSerde,
 };
 
+/*
+ * Shared file-writing helpers for CSS and codegen outputs.
+ */
 #[wasm_bindgen]
 impl WasmCompiler {
-    fn write_relative_files<'a>(
+    pub(super) fn write_relative_files<'a>(
         &self,
         root: &str,
         files: impl IntoIterator<Item = (&'a str, &'a str)>,
@@ -37,28 +43,7 @@ impl WasmCompiler {
         Ok(written)
     }
 
-    fn write_artifacts_to_root(
-        &self,
-        root: &str,
-        artifacts: &[CodegenArtifactSerde],
-    ) -> Result<Vec<String>, JsValue> {
-        let mut written = Vec::new();
-        for artifact in artifacts {
-            written.extend(
-                self.write_relative_files(
-                    root,
-                    artifact
-                        .files
-                        .iter()
-                        .map(|file| (file.path.as_str(), file.code.as_str())),
-                    "artifact",
-                )?,
-            );
-        }
-        Ok(written)
-    }
-
-    fn write_target_file(&self, target: &str, code: &str) -> Result<(), JsValue> {
+    pub(super) fn write_target_file(&self, target: &str, code: &str) -> Result<(), JsValue> {
         let parent = self.paths.dirname(target);
         if !parent.is_empty() {
             self.fs
@@ -71,40 +56,9 @@ impl WasmCompiler {
             .map_err(|err| JsValue::from_str(&err.to_string()))
     }
 
-    /// Generate artifacts and write them under `outdir` via the in-memory fs.
-    /// Returns the written paths.
-    ///
-    /// # Errors
-    /// Returns a JS error if a file fails to write or results fail to serialize.
-    #[wasm_bindgen(js_name = writeArtifacts)]
-    pub fn write_artifacts(&self, options: JsValue) -> Result<JsValue, JsValue> {
-        let options = write_artifacts_options_from_js(options)?;
-        let artifacts = if let Some(artifacts) = options.artifacts {
-            artifacts
-        } else {
-            self.inner
-                .generate_artifacts(
-                    &self.user_config,
-                    generate_options(
-                        &self.user_config,
-                        &GenerateArtifactOptionsSerde {
-                            force_import_extension: options.force_import_extension,
-                        },
-                    ),
-                )
-                .into_iter()
-                .map(to_codegen_artifact)
-                .collect()
-        };
-        let cwd = options.cwd.unwrap_or_else(|| self.user_config.cwd.clone());
-        let base = self.paths.resolve(&cwd, &options.outdir);
-        let written = self.write_artifacts_to_root(&base, &artifacts)?;
-        let serializer = serde_wasm_bindgen::Serializer::new().serialize_maps_as_objects(true);
-        written
-            .serialize(&serializer)
-            .map_err(|err| JsValue::from_str(&err.to_string()))
-    }
-
+    /*
+     * CSS compile/write entrypoints.
+     */
     /// Compile to CSS. Mirrors the NAPI `Project.compile()`.
     ///
     /// # Errors
@@ -235,5 +189,228 @@ impl WasmCompiler {
         result
             .serialize(&serializer)
             .map_err(|err| JsValue::from_str(&err.to_string()))
+    }
+}
+
+/*
+ * CSS output assembly.
+ */
+fn build_compile_output(
+    project: &mut pandacss_project::Project,
+    user_config: &pandacss_config::UserConfig,
+    static_pattern_atoms: &[CoreAtom],
+    static_pattern_diagnostics: Vec<pandacss_extractor::Diagnostic>,
+    emit_layer_declaration: bool,
+    minify_override: Option<bool>,
+) -> CompileOutputSerde {
+    let token_dictionary = project.config().token_dictionary();
+    let manifest = compile_manifest_serde(project, token_dictionary.as_ref());
+    let output = build_stylesheet_output(
+        project,
+        user_config,
+        token_dictionary,
+        static_pattern_atoms,
+        emit_layer_declaration,
+        minify_override,
+    );
+    let diagnostics = project
+        .diagnostics()
+        .iter()
+        .cloned()
+        .chain(static_pattern_diagnostics)
+        .chain(output.diagnostics)
+        .collect();
+    CompileOutputSerde {
+        css: output.css,
+        source_map: output.source_map,
+        manifest,
+        layer_ranges: layer_ranges_from(&output.layer_ranges),
+        diagnostics,
+    }
+}
+
+fn build_layer_compile_output(
+    project: &mut pandacss_project::Project,
+    user_config: &pandacss_config::UserConfig,
+    static_pattern_atoms: &[CoreAtom],
+    static_pattern_diagnostics: Vec<pandacss_extractor::Diagnostic>,
+    options: &LayerCssOptionsSerde,
+) -> CompileOutputSerde {
+    let token_dictionary = project.config().token_dictionary();
+    let manifest = compile_manifest_serde(project, token_dictionary.as_ref());
+    let output = build_stylesheet_output(
+        project,
+        user_config,
+        token_dictionary,
+        static_pattern_atoms,
+        false,
+        options.minify,
+    );
+    let selected: Vec<_> = options
+        .layers
+        .iter()
+        .filter_map(|name| pandacss_stylesheet::StylesheetLayer::from_name(name))
+        .collect();
+    let mut css = output.get_layer_css(&selected);
+    if options.emit_layer_declaration.unwrap_or(false) {
+        let preamble =
+            pandacss_stylesheet::layer_order_declaration(&user_config.layers, Some(&selected));
+        if !preamble.is_empty() {
+            css.insert_str(0, &format!("{preamble}\n"));
+        }
+    }
+    let diagnostics = project
+        .diagnostics()
+        .iter()
+        .cloned()
+        .chain(static_pattern_diagnostics)
+        .chain(output.diagnostics)
+        .collect();
+    CompileOutputSerde {
+        css,
+        source_map: output.source_map,
+        manifest,
+        layer_ranges: empty_layer_ranges(),
+        diagnostics,
+    }
+}
+
+fn build_split_css(
+    project: &mut pandacss_project::Project,
+    user_config: &pandacss_config::UserConfig,
+    static_pattern_atoms: &[CoreAtom],
+    css_options: &CssOutputOptionsSerde,
+) -> Vec<SplitCssFileSerde> {
+    let token_dictionary = project.config().token_dictionary();
+    let snapshots = project.stylesheet_snapshots(user_config);
+    let selected_layers = css_options.layers.as_ref().map(|layers| {
+        layers
+            .iter()
+            .filter_map(|name| pandacss_stylesheet::StylesheetLayer::from_name(name))
+            .collect::<Vec<_>>()
+    });
+    let options = pandacss_stylesheet::StylesheetOptions {
+        minify: resolve_minify(user_config, css_options.minify),
+        include_static: pandacss_stylesheet::has_static_css(user_config),
+        source_map: false,
+        emit_layer_declaration: css_options.emit_layer_declaration.unwrap_or(true),
+        layers: selected_layers,
+    };
+    pandacss_stylesheet::split_css(
+        &pandacss_stylesheet::StylesheetInput {
+            config: user_config,
+            token_dictionary,
+            atoms: snapshots.atoms,
+            utility_styles: snapshots.utility_styles,
+            encoded_recipes: snapshots.encoded_recipes,
+            static_encoded_recipes: Some(snapshots.static_encoded_recipes),
+            static_pattern_atoms,
+            token_refs: snapshots.token_refs,
+        },
+        &options,
+    )
+    .into_iter()
+    .map(|file| SplitCssFileSerde {
+        path: file.path,
+        code: file.code,
+    })
+    .collect()
+}
+
+fn build_stylesheet_output(
+    project: &mut pandacss_project::Project,
+    user_config: &pandacss_config::UserConfig,
+    token_dictionary: Option<Arc<pandacss_tokens::TokenDictionary>>,
+    static_pattern_atoms: &[CoreAtom],
+    emit_layer_declaration: bool,
+    minify_override: Option<bool>,
+) -> pandacss_stylesheet::StylesheetOutput {
+    let snapshots = project.stylesheet_snapshots(user_config);
+    let options = pandacss_stylesheet::StylesheetOptions {
+        minify: resolve_minify(user_config, minify_override),
+        include_static: pandacss_stylesheet::has_static_css(user_config),
+        source_map: false,
+        emit_layer_declaration,
+        layers: None,
+    };
+    pandacss_stylesheet::compile(
+        pandacss_stylesheet::StylesheetInput {
+            config: user_config,
+            token_dictionary,
+            atoms: snapshots.atoms,
+            utility_styles: snapshots.utility_styles,
+            encoded_recipes: snapshots.encoded_recipes,
+            static_encoded_recipes: Some(snapshots.static_encoded_recipes),
+            static_pattern_atoms,
+            token_refs: snapshots.token_refs,
+        },
+        &options,
+    )
+}
+
+fn resolve_minify(
+    user_config: &pandacss_config::UserConfig,
+    minify_override: Option<bool>,
+) -> bool {
+    minify_override.unwrap_or_else(|| {
+        user_config
+            .extra
+            .get("minify")
+            .and_then(serde_json::Value::as_bool)
+            .unwrap_or(false)
+    })
+}
+
+/*
+ * Compile output serialization.
+ */
+fn compile_manifest_serde(
+    project: &pandacss_project::Project,
+    token_dictionary: Option<&Arc<pandacss_tokens::TokenDictionary>>,
+) -> CompileManifestSerde {
+    let files = project
+        .file_manifest()
+        .into_iter()
+        .map(|(path, hash)| CompileFileManifestSerde {
+            path: path.as_ref().to_owned(),
+            hash: format!("{hash:016x}"),
+        })
+        .collect();
+    let tokens = token_dictionary.map_or_else(Vec::new, |dict| {
+        let mut paths = BTreeSet::new();
+        for token in dict.iter() {
+            paths.insert(token.path.to_string());
+        }
+        paths.into_iter().collect()
+    });
+    CompileManifestSerde { files, tokens }
+}
+
+fn empty_layer_ranges() -> CompileLayerRangesSerde {
+    CompileLayerRangesSerde {
+        reset: None,
+        base: None,
+        tokens: None,
+        recipes: None,
+        utilities: None,
+    }
+}
+
+fn layer_ranges_from(
+    ranges: &pandacss_stylesheet::StylesheetLayerRanges,
+) -> CompileLayerRangesSerde {
+    CompileLayerRangesSerde {
+        reset: ranges.reset.as_ref().map(to_serde_range),
+        base: ranges.base.as_ref().map(to_serde_range),
+        tokens: ranges.tokens.as_ref().map(to_serde_range),
+        recipes: ranges.recipes.as_ref().map(to_serde_range),
+        utilities: ranges.utilities.as_ref().map(to_serde_range),
+    }
+}
+
+fn to_serde_range(range: &std::ops::Range<usize>) -> CompileLayerRangeSerde {
+    CompileLayerRangeSerde {
+        start: u32::try_from(range.start).unwrap_or(u32::MAX),
+        end: u32::try_from(range.end).unwrap_or(u32::MAX),
     }
 }
