+++ packages/compiler-wasm/crate/src/project/recipes.rs
@@ -5,8 +5,8 @@ use pandacss_extractor::Literal;
 use serde::Serialize as _;
 use wasm_bindgen::prelude::*;
 
+use super::interop::{collect_sorted_atoms, slice_to_atom_serde};
 use super::serde_types::{RecipeEntrySerde, StaticPatternResultSerde};
-use super::support::{collect_sorted_atoms, slice_to_atom_serde};
 use super::transforms::apply_pattern_transform;
 
 #[wasm_bindgen]
